fixed ui fps, button hover and systems
and added a fix for intra-state mouse clicking.

+++ header/Components/UIComponents/UIGameLevelComponent.h
@@ -7,7 +7,7 @@
 
 struct UIGameLevelComponent
 {
-	float value;
+	int value;
 	DirectX::XMFLOAT2 position;
 	DirectX::XMFLOAT2 scale;
 	UIImage image;
